+++ Runtime/Internal/UnityOpenXRNativeApi.cs
@@ -38,13 +38,6 @@ namespace Google.XR.Extensions.Internal
                 enabled, result >= 0  ? "succeed" : "failed", result);
         }
 
-        public static void MetaSetSpaceWarp(bool enabled)
-        {
-            int result = ExternalApi.MetaSetSpaceWarp(enabled);
-            Debug.LogFormat("MetaSetSpaceWarp({0}): {1} with status {2}.",
-                enabled, result >= 0 ? "succeed" : "failed", result);
-        }
-
         public static void FBSetFoveationLevel(
             ulong xrSession, XRFoveationLevel foveationLevel,
             float verticalOffset, bool foveationDynamic)
@@ -60,9 +53,6 @@ namespace Google.XR.Extensions.Internal
         {
             [DllImport(ApiConstants.UnityOpenXRLib)]
             public static extern int MetaSetSubsampledLayout(bool enabled);
-
-            [DllImport(ApiConstants.UnityOpenXRLib)]
-            public static extern int MetaSetSpaceWarp(bool enable);
 
             [DllImport(ApiConstants.UnityOpenXRLib)]
             public static extern int FBSetFoveationLevel(